+++ Samples/iOS-ObjectiveC/iOS-ObjectiveC/AppDelegate.m
@@ -16,8 +16,6 @@ AppDelegate ()
     [SentrySDK startWithConfigureOptions:^(SentryOptions *options) {
         options.dsn = @"https://a92d50327ac74b8b9aa4ea80eccfb267@o447951.ingest.sentry.io/5428557";
         options.debug = YES;
-        options.logLevel = kSentryLogLevelVerbose;
-        options.attachStacktrace = YES;
         options.sessionTrackingIntervalMillis = [@5000 unsignedIntegerValue];
     }];