+++ Samples/iOS-Swift/iOS-Swift/ViewController.swift
@@ -24,7 +24,7 @@ class ViewController: UIViewController {
         }
         // Also works
         let user = Sentry.User(userId: "1")
-        user.email = "tony@example.com"
+        user.email = "tony1@example.com"
         SentrySDK.setUser(user)
     }
     
@@ -80,6 +80,13 @@ class ViewController: UIViewController {
         SentrySDK.capture(exception: exception, scope: scope)
     }
     
+    @IBAction func captureTransaction(_ sender: Any) {
+        let transaction = SentrySDK.startTransaction(name: "Some Transaction", operation: "Some Operation")
+        DispatchQueue.main.asyncAfter(deadline: .now() + Double.random(in: 0.4...0.6), execute: {
+            transaction.finish()
+        })
+    }
+   
     @IBAction func crash(_ sender: Any) {
         SentrySDK.crash()
     }