+++ Sources/Sentry/SentrySDK.m
@@ -8,8 +8,6 @@
 #import "SentryMeta.h"
 #import "SentryScope.h"
 
-static SentryLogLevel logLevel = kSentryLogLevelError;
-
 @interface
 SentrySDK ()
 
@@ -25,8 +23,6 @@ NS_ASSUME_NONNULL_BEGIN
 
 static SentryHub *currentHub;
 static BOOL crashedLastRunCalled;
-
-@dynamic logLevel;
 
 + (SentryHub *)currentHub
 {
@@ -61,10 +57,9 @@ static BOOL crashedLastRunCalled;
     SentryOptions *options = [[SentryOptions alloc] initWithDict:optionsDict
                                                 didFailWithError:&error];
     if (nil != error) {
-        [SentryLog logWithMessage:@"Error while initializing the SDK"
-                         andLevel:kSentryLogLevelError];
+        [SentryLog logWithMessage:@"Error while initializing the SDK" andLevel:kSentryLevelError];
         [SentryLog logWithMessage:[NSString stringWithFormat:@"%@", error]
-                         andLevel:kSentryLogLevelError];
+                         andLevel:kSentryLevelError];
     } else {
         [SentrySDK startWithOptionsObject:options];
     }
@@ -72,14 +67,14 @@ static BOOL crashedLastRunCalled;
 
 + (void)startWithOptionsObject:(SentryOptions *)options
 {
-    [self setLogLevel:options.logLevel];
+    [SentryLog configure:options.debug diagnosticLevel:options.diagnosticLevel];
     SentryClient *newClient = [[SentryClient alloc] initWithOptions:options];
     // The Hub needs to be initialized with a client so that closing a session
     // can happen.
     [SentrySDK setCurrentHub:[[SentryHub alloc] initWithClient:newClient andScope:nil]];
     [SentryLog logWithMessage:[NSString stringWithFormat:@"SDK initialized! Version: %@",
                                         SentryMeta.versionString]
-                     andLevel:kSentryLogLevelDebug];
+                     andLevel:kSentryLevelDebug];
     [SentrySDK installIntegrations];
 }
 
@@ -97,12 +92,12 @@ static BOOL crashedLastRunCalled;
 
 + (SentryId *)captureEvent:(SentryEvent *)event
 {
-    return [SentrySDK captureEvent:event withScope:[SentrySDK.currentHub getScope]];
+    return [SentrySDK captureEvent:event withScope:SentrySDK.currentHub.scope];
 }
 
 + (SentryId *)captureEvent:(SentryEvent *)event withScopeBlock:(void (^)(SentryScope *))block
 {
-    SentryScope *scope = [[SentryScope alloc] initWithScope:[SentrySDK.currentHub getScope]];
+    SentryScope *scope = [[SentryScope alloc] initWithScope:SentrySDK.currentHub.scope];
     block(scope);
     return [SentrySDK captureEvent:event withScope:scope];
 }
@@ -112,14 +107,24 @@ static BOOL crashedLastRunCalled;
     return [SentrySDK.currentHub captureEvent:event withScope:scope];
 }
 
++ (id<SentrySpan>)startTransactionWithName:(NSString *)name operation:(NSString *)operation
+{
+    return [SentrySDK.currentHub startTransactionWithName:name operation:operation];
+}
+
++ (id<SentrySpan>)startTransactionWithContext:(SentryTransactionContext *)transactionContext
+{
+    return [SentrySDK.currentHub startTransactionWithContext:transactionContext];
+}
+
 + (SentryId *)captureError:(NSError *)error
 {
-    return [SentrySDK captureError:error withScope:[SentrySDK.currentHub getScope]];
+    return [SentrySDK captureError:error withScope:SentrySDK.currentHub.scope];
 }
 
 + (SentryId *)captureError:(NSError *)error withScopeBlock:(void (^)(SentryScope *_Nonnull))block
 {
-    SentryScope *scope = [[SentryScope alloc] initWithScope:[SentrySDK.currentHub getScope]];
+    SentryScope *scope = [[SentryScope alloc] initWithScope:SentrySDK.currentHub.scope];
     block(scope);
     return [SentrySDK captureError:error withScope:scope];
 }
@@ -131,13 +136,13 @@ static BOOL crashedLastRunCalled;
 
 + (SentryId *)captureException:(NSException *)exception
 {
-    return [SentrySDK captureException:exception withScope:[SentrySDK.currentHub getScope]];
+    return [SentrySDK captureException:exception withScope:SentrySDK.currentHub.scope];
 }
 
 + (SentryId *)captureException:(NSException *)exception
                 withScopeBlock:(void (^)(SentryScope *))block
 {
-    SentryScope *scope = [[SentryScope alloc] initWithScope:[SentrySDK.currentHub getScope]];
+    SentryScope *scope = [[SentryScope alloc] initWithScope:SentrySDK.currentHub.scope];
     block(scope);
     return [SentrySDK captureException:exception withScope:scope];
 }
@@ -149,12 +154,12 @@ static BOOL crashedLastRunCalled;
 
 + (SentryId *)captureMessage:(NSString *)message
 {
-    return [SentrySDK captureMessage:message withScope:[SentrySDK.currentHub getScope]];
+    return [SentrySDK captureMessage:message withScope:SentrySDK.currentHub.scope];
 }
 
 + (SentryId *)captureMessage:(NSString *)message withScopeBlock:(void (^)(SentryScope *))block
 {
-    SentryScope *scope = [[SentryScope alloc] initWithScope:[SentrySDK.currentHub getScope]];
+    SentryScope *scope = [[SentryScope alloc] initWithScope:SentrySDK.currentHub.scope];
     block(scope);
     return [SentrySDK captureMessage:message withScope:scope];
 }
@@ -177,17 +182,6 @@ static BOOL crashedLastRunCalled;
 + (void)configureScope:(void (^)(SentryScope *scope))callback
 {
     [SentrySDK.currentHub configureScope:callback];
-}
-
-+ (void)setLogLevel:(SentryLogLevel)level
-{
-    NSParameterAssert(level);
-    logLevel = level;
-}
-
-+ (SentryLogLevel)logLevel
-{
-    return logLevel;
 }
 
 /**
@@ -233,21 +227,21 @@ static BOOL crashedLastRunCalled;
             NSString *logMessage = [NSString stringWithFormat:@"[SentryHub doInstallIntegrations] "
                                                               @"couldn't find \"%@\" -> skipping.",
                                              integrationName];
-            [SentryLog logWithMessage:logMessage andLevel:kSentryLogLevelError];
+            [SentryLog logWithMessage:logMessage andLevel:kSentryLevelError];
             continue;
         } else if ([SentrySDK.currentHub isIntegrationInstalled:integrationClass]) {
             NSString *logMessage =
                 [NSString stringWithFormat:@"[SentryHub doInstallIntegrations] already "
                                            @"installed \"%@\" -> skipping.",
                           integrationName];
-            [SentryLog logWithMessage:logMessage andLevel:kSentryLogLevelError];
+            [SentryLog logWithMessage:logMessage andLevel:kSentryLevelError];
             continue;
         }
         id<SentryIntegrationProtocol> integrationInstance = [[integrationClass alloc] init];
         [integrationInstance installWithOptions:options];
         [SentryLog
             logWithMessage:[NSString stringWithFormat:@"Integration installed: %@", integrationName]
-                  andLevel:kSentryLogLevelDebug];
+                  andLevel:kSentryLevelDebug];
         [SentrySDK.currentHub.installedIntegrations addObject:integrationInstance];
     }
 }