+++ Tests/SentryTests/SentryClientTests.swift
@@ -3,6 +3,9 @@
 import XCTest
 
 class SentryClientTest: XCTestCase {
+    
+    private static let dsnAsString = TestConstants.dsnAsString(username: "SentryClientTest")
+    private static let dsn = TestConstants.dsn(username: "SentryClientTest")
 
     private class Fixture {
         let transport = TestTransport()
@@ -38,14 +41,14 @@ class SentryClientTest: XCTestCase {
             user.email = "someone@sentry.io"
             user.ipAddress = "127.0.0.1"
             
-            fileManager = try! SentryFileManager(dsn: TestConstants.dsn, andCurrentDateProvider: TestCurrentDateProvider())
+            fileManager = try! SentryFileManager(dsn: SentryClientTest.dsn, andCurrentDateProvider: TestCurrentDateProvider())
         }
 
         func getSut(configureOptions: (Options) -> Void = { _ in }) -> Client {
             var client: Client!
             do {
                 let options = try Options(dict: [
-                    "dsn": TestConstants.dsnAsString
+                    "dsn": SentryClientTest.dsnAsString
                 ])
                 configureOptions(options)
 
@@ -157,7 +160,7 @@ class SentryClientTest: XCTestCase {
             }
         }
     }
-    
+      
     func testCaptureEventWithException() {
         let event = Event()
         event.exceptions = [ Exception(value: "", type: "")]
@@ -187,7 +190,7 @@ class SentryClientTest: XCTestCase {
             options.dsn = nil
         })
         
-        sut.options.dsn = TestConstants.dsnAsString
+        sut.options.dsn = SentryClientTest.dsnAsString
         
         let eventId = sut.capture(event: event)
         eventId.assertIsNotEmpty()
@@ -622,7 +625,7 @@ class SentryClientTest: XCTestCase {
         SentryFileManager.prepareInitError()
         
         let options = Options()
-        options.dsn = TestConstants.dsnAsString
+        options.dsn = SentryClientTest.dsnAsString
         let client = Client(options: options)
         
         XCTAssertNil(client)